avoid initially creating an empty resource
This fixes a bug where, the first time a set is harvested, none of the
resource metadata gets indexed in solr. Harvesting the same set again
would populate the data in solr.

My best guess as to what was happening is that #find_or_create_by was
creating an empty sidecar doc, which then for some reason wasn't getting
updated when calling #save_and_index on its resource.

Instead, I've opted to set up all the data before committing anything to
the db/solr. Now, the first harvest of a set brings in the metadata as
expected.