i133 Ensure all data updates if an item has already been harvested (#22)
* ensure all data updates if an item has already been harvested

* Changing resource reload to not overwrite the changes we are trying to make

* use #save! to throw error, fix typo

* minor clarity improvement

* avoid initially creating an empty resource

This fixes a bug where, the first time a set is harvested, none of the
resource metadata gets indexed in solr. Harvesting the same set again
would populate the data in solr.

My best guess as to what was happening is that #find_or_create_by was
creating an empty sidecar doc, which then for some reason wasn't getting
updated when calling #save_and_index on its resource.

Instead, I've opted to set up all the data before committing anything to
the db/solr. Now, the first harvest of a set brings in the metadata as
expected.

Co-authored-by: Maura C <maura_carbone@harvard.edu>