Removing Spotlight::HarvestingCompleteMailer.harvest_failed UI call
Given that we're uploading via the UI, we don't need to send an email
when we fail the save of the OaipmhHarvester object.  Instead let's
flash the errors and not send an email.

Then if the harvester saves, when we run the
`Spotlight::Resources::PerformHarvestsJob.perform_later` that can send
the email notifying of error.

See https://github.com/harvard-lts/CURIOSity/issues/185