Adding more robust reporting for PerformHarvestJobs
Prior to this commit we had two emails that were sent:

1. **harvest_failed**: Regardless of `Harvester#harvest_oai_items`
   results, if there were sidecar items missing, we'd send an
   `harvest_failed` email.  This email did not include any additional
   information on what sidecars were missing.
2. **harvest_succeeded**: If all harvest documents had existing
   sidecars, we'd send a `harvest_succeeded` email.

*Note:* prior to and including this change, there's not handling of errors
encountered elsewhere (e.g. invalid data, network issues, database
failures, etc).

With this commit, we consolidate the `harvest_failed` and
`harvest_succeeded` into `harvest_set_completed`.  That
`harvest_set_completed` means that the set completed and there may or
may not be issues with document's related sidecars.  If there are
issues, we include details of those issues in the email.

Closes #185