only perform HTTP request if there is only one uri
Some records currently get parsed with multiple thumbnail/full image urls concatenated together. This was throwing errors and preventing the records from coming in at all.
For parity with the old site, we now allow these records to be imported with broken images so that at least the record is in the system.