Update curiosity.yml
Added new xpath to additional-digital-items_tesim for BTA. If this is the preferred label, we can delete source-institution-digitization_tesim. source-institution-digitization_tesim currently has label as well which is better for accessibility, but the link is not hyperlinked in CURIOSity. Need to confirm with BTA which label and display is preferred.