Add index for `iiif_manifest_url_ssi` (#81)
* Update curiosity.yml

* Add index for `iiif_manifest_url_ssi`

This commit will add an index field for `iiif_manifest_url_ssi` which
would be consumed by Spotlight's appearances controller to set images
(like thumbnails and mastheads) based on the IIIF Manifest of the
resource.  Because Harvard's URIs redirect, a proxy endpoint from the
server is needed to receive the JSON from the redirect.  This is
assuming the route at `iiif/proxy` with the param url=<url>.  This is
what we will index onto the resource to get the IIIF Manifest picked up
by Spotlight.

---------

Co-authored-by: dl-maura <maura_carbone@harvard.edu>
Co-authored-by: Vanessa Venti <60190174+VanessaVenti@users.noreply.github.com>